+++ packages/app/src/services/apiClient/fetcher.ts
@@ -1,7 +1,7 @@
 import { ApiError } from "next/dist/server/api-utils";
 
-import type { Any } from "../../common/utils/types";
 import { useUserStore } from "./useUser";
+import type { Any } from "@common/utils/types";
 
 const API_URL = process.env.NEXT_PUBLIC_API_URL;
