+++ packages/app/src/services/apiClient/useMe.ts
@@ -1,4 +1,5 @@
 import useSWR from "swr";
+
 import { fetcher } from "./fetcher";
 
 // TODO: type à confirmer par rapport au endpoint /!\
