+++ packages/app/src/services/apiClient/useStats.ts
@@ -1,8 +1,8 @@
 import useSWR from "swr";
 
-import { buildUrlParamsString } from "../../common/utils/url";
-import { fetcher } from "../../services/apiClient";
+import { buildUrlParamsString } from "@common/utils/url";
 import type { FetcherReturn } from "@services/apiClient";
+import { fetcher } from "@services/apiClient";
 
 export type StatsType = {
   avg: number;
